requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
asttokens=3.0.0=pyhd8ed1ab_1
blas=1.0=mkl
bottleneck=1.4.2=py312h4b0e54e_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py312hd77b12b_8
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.11.26=haa95532_0
certifi=2024.8.30=py312haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=py312h214f63a_0
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=12.4.127=0
cuda-cudart-dev=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-libraries-dev=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvrtc-dev=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-opencl-dev=12.6.77=0
cuda-profiler-api=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.8.10=py312h275cf98_0
decorator=5.1.1=pyhd8ed1ab_1
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=he0c23c2_0
filelock=3.13.1=py312haa95532_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=ha860e81_0
gettext=0.22.5=h5728263_3
gettext-tools=0.22.5=h5a7288d_3
giflib=5.2.2=h7edc060_0
glib=2.80.2=h0df6a38_0
glib-tools=2.80.2=h2f9d560_0
gst-plugins-base=1.18.5=he07aa86_3
gstreamer=1.18.5=hdff456e_3
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.7=py312haa95532_0
importlib-metadata=8.5.0=pyha770c72_1
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.30.0=pyh7428d3b_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=py312haa95532_1
joblib=1.4.2=py312haa95532_0
jpeg=9e=h827c3e9_3
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=py312h2e8e312_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.21.3=hdf4eb48_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libasprintf=0.22.5=h5728263_3
libasprintf-devel=0.22.5=h5728263_3
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=12.0.1=default_h81446c8_4
libclang13=14.0.6=default_h8e68704_1
libcublas=12.4.5.8=0
libcublas-dev=12.4.5.8=0
libcufft=11.2.1.3=0
libcufft-dev=11.2.1.3=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.6.1.9=0
libcusolver-dev=11.6.1.9=0
libcusparse=12.3.1.170=0
libcusparse-dev=12.3.1.170=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.6.4=he0c23c2_0
libffi=3.4.4=hd77b12b_1
libgettextpo=0.22.5=h5728263_3
libgettextpo-devel=0.22.5=h5728263_3
libglib=2.80.2=h0df6a38_0
libiconv=1.17=hcfcfb64_2
libintl=0.22.5=h5728263_3
libintl-devel=0.22.5=h5728263_3
libjpeg-turbo=2.0.0=h196d8e1_0
liblzma=5.6.3=h2466b09_1
liblzma-devel=5.6.3=h2466b09_1
libnpp=12.2.5.30=0
libnpp-dev=12.2.5.30=0
libnvfatbin=12.6.77=0
libnvfatbin-dev=12.6.77=0
libnvjitlink=12.4.127=0
libnvjitlink-dev=12.4.127=0
libnvjpeg=12.3.1.117=0
libnvjpeg-dev=12.3.1.117=0
libogg=1.3.5=h2466b09_0
libpng=1.6.39=h8cc25b3_0
libpq=16.6=h5b74ccd_1
libsodium=1.0.20=hc70643c_0
libsqlite=3.47.2=h67fdade_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
libzlib=1.2.13=h2466b09_6
lightning-utilities=0.9.0=py312haa95532_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py312h2bbff1b_0
matplotlib=3.9.2=py312haa95532_1
matplotlib-base=3.9.2=py312he19b0ae_1
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.11=py312h827c3e9_0
mkl_random=1.2.8=py312h0158946_0
mlxtend=0.23.1=py312haa95532_0
mpmath=1.3.0=py312haa95532_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.2.1=py312haa95532_0
numexpr=2.10.1=py312h4cd664f_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
openjpeg=2.5.2=hae555c5_0
openssl=3.4.0=h2466b09_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312h5da7b33_0
parkinson-disease-hand-writing-detection=0.1=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.43=h17e33f8_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=py312hb5480e2_0
pip=24.3.1=pyh8b19718_0
platformdirs=4.3.6=pyhd8ed1ab_1
ply=3.11=py312haa95532_1
prompt-toolkit=3.0.48=pyha770c72_1
psutil=6.1.0=py312h4389bb4_0
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.18.0=pyhd8ed1ab_1
pyparsing=3.2.0=py312haa95532_0
pyqt=5.15.10=py312hd77b12b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
pysocks=1.7.1=py312haa95532_0
python=3.12.3=h2628c8c_0_cpython
python-dateutil=2.9.0post0=py312haa95532_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytorch=2.5.1=py3.12_cuda12.4_cudnn9_0
pytorch-cuda=12.4=h3fd98bf_7
pytorch-mutex=1.0=cuda
pytz=2024.1=py312haa95532_0
pywin32=307=py312h275cf98_3
pyyaml=6.0.2=py312h827c3e9_0
pyzmq=26.2.0=py312hd7027bb_3
qt-main=5.15.2=he8e5bd7_7
requests=2.32.3=py312haa95532_1
scikit-learn=1.5.1=py312h0158946_0
scipy=1.14.1=py312h9d85e7c_0
setuptools=75.6.0=pyhff2d567_1
sip=6.7.12=py312hd77b12b_0
six=1.17.0=pyhd8ed1ab_0
sqlite=3.47.2=h2466b09_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.2=py312haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.5.0=py312hfc267ef_0
tk=8.6.14=h0416ee5_0
torchaudio=2.5.1=pypi_0
torchinfo=1.8.0=pyhd8ed1ab_0
torchmetrics=1.4.0.post0=py312haa95532_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=py312h4389bb4_0
tqdm=4.66.5=py312hfc267ef_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=hc8b5060_0
ucrt=10.0.22621.0=h57928b3_1
unicodedata2=15.1.0=py312h2bbff1b_0
urllib3=2.2.3=py312haa95532_0
vc=14.40=haa95532_2
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
win_inet_pton=1.1.0=py312haa95532_0
xz=5.6.3=h208afaa_1
xz-tools=5.6.3=h2466b09_1
yaml=0.2.5=he774522_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h2466b09_6
zstd=1.5.6=h8880b57_0
